CP-2138: License daemon should store cut-off date of last held license
We add a new function to lpe.ml called write_sa_date that extracts the
SA date from the LPE_LPE.ini file. In the future this function will
call the LPE API directly to get the SA date, but this API call
doesn't exist yet.

This function is called: 1) when the LPE is shut down, 2) when we
change editions under the same license (same as the first case), 3)
when we apply a new license, and 4) when we shutdown v6d cleanly. We
probably only need the last case, but the first three cases seem
reasonable as well.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>