PR-1365: XenServer for Kaviza license entitlement
Kaviza is Citrix's new virtual desktop product targeted at the SMB, or user
groups of 300 users or less.  We would like to support Kaviza in much the same
was as we do XenDesktop - by giving Kaviza customers the entitlement to run
XenServer Advanced Edition for their "VDI in a box" infrastructure.

In order to do this before the Boston release, we will be giving Kaviza
customers access to the Enterprise edition, in the same way that we do for
XenDesktop users. This change will allow users with a Kaviza license to apply
the XenServer Enterprise for XenDesktop edition, without having a XenDesktop
license. In the future, we'll look at creating a XenServer Advanced for Kaviza
edition.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>