CP-2676: Egenera license build
This patch contains some major refactoring in order to better support the
creation of new flavors of the v6 daemon. We have changed the License,
License_file and Realv6 modules into functors which are parameterized on the
Editions module type. This way, we can create new Editions modules for each v6d
flavor, and then drop them into Realv6.Make in order to create a new daemon.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>