CA-72736: Assume license permanent if expires after 50 years
It appears that if a license has both 1) a "permanent" expiry, and 2) an
annoyance profile (something that Ring3 hasn't seen before), then the LPE
returns a days_to_expiry value of 1000 years in the future. The fix is that we
now assume that any value greater than 50 years is a permanent license.

Mike McClurg <mike.mcclurg@citrix.com>