CA-81717: Upgrade from GA to EAP w/ no TP license gives license expired
On first boot, v6d is passed the current license parameters, which don't
contain the early_release parameter. This puts v6d into a weird split state: it
knows to look for a TP (tech preview) license on the license server, but it
doesn't know that the early_release parameter should be set. So it asks the
license server for a beta license, doesn't get one, and then doesn't know that
it should put XenServer in an upgrade-grace state. This function will return
the new license parameters, and will then correctly set the early_release
parameter. So the second time you call apply_edition, it will then fail to
retrieve a beta license, but will know that it should be in the upgrade-grace
state.

We now check for the V6globs.beta boolean, as well as checking that we have the
early_release key in the current xapi database.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>