CA-83054: Fix expired-becomes-permanent bug
We used to use a days_to_expiry value of (-1) to indicate that a license never
expires. Setting the date forward beyond the expiry date, and then restarting
xapi, caused v6d to assume that a valid license was still checked out, and that
the expiry date was negative. This caused us to assume that the license was
checked out, valid, and permanent, when in fact it had been expired.

We now use the type Lpe.expiry_t for all expiry dates, so that we have can be
sure whether a license is permanent or not.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>