CA-103762: v6d returns an 'effective edition' on startup when license unavailable
Since we no longer restrive VM.start in xapi when a license expires, it no
longer makes sense to just expire the license and leave the set of restricted
features unchanged. v6d now returns an "effective edition" on startup, which is
the free edition if the requested edition is unavailable.

Perhaps a better way to do this in the future would be to move some additional
feature flags to exist only in xapi, not v6d, so that xapi can decide whether
to restrict these or not, depending on license expiry state.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>